The nodeContainers Array is an Object
I wanted the access key to each nodeContainer to be that node's ID to
simplify ID systems. I believe the best way to do this is to make the
object that contains all those containers be an object whose keys
correspond to the node containers whose node's ID matches.

I'm going to test it. I might go to an array with gaps or scrap it all.